+++ test/integration/telemetry_test.exs
@@ -8,7 +8,7 @@ defmodule Puck.Integration.TelemetryTest do
   describe "BAML telemetry" do
     @describetag :baml
 
-    setup :check_ollama_available!
+    setup :check_fireworks_available!
 
     setup do
       client =